chore: clean up dead code and lint warnings (no behavior change)
Apply machine-applicable rustc + clippy fixes and remove confirmed
no-op code introduced by earlier edits. Verified: full release-mode
build succeeds and the binary runs; warnings reduced 1229 -> 551.

- Remove unused imports across frameworks/libc modules
- Remove 29 dead `drop(host)` no-ops (host is a plain reference, so
  drop() does nothing and does not affect borrow lifetimes)
- Drop unnecessary same-type casts, convert manual assign ops to +=/etc.,
  use is_multiple_of() where suggested

No functional/behavioral changes; intentional HLE stubs (e.g. OpenGL
no-op functions) and single-threaded `static mut` state left untouched.

+++ src/frameworks/foundation/ns_value.rs
@@ -632,7 +632,7 @@ pub const CLASSES: ClassExports = objc_classes! {
             coder,
             class
         );
-        msg_class![env; NSNumber numberWithInt:(0 as i32)]
+        msg_class![env; NSNumber numberWithInt:(0_i32)]
     };
     release(env, this);
     new_num
